requirements.txt
Flask==2.3.3
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
